add object headers
