+++ src/edge.hpp
@@ -1,0 +1,8 @@
+#ifndef h_edge
+#define h_edge
+
+class Edge{
+    int m_direction;
+};
+
+#endif
