+++ src/point.hpp
@@ -1,0 +1,10 @@
+#ifndef h_point
+#define h_point
+
+class Point{
+    double m_x;
+    double m_y;
+    int m_id;
+};
+
+#endif
